get boilerplate graphql running
