+++ api/package.json
@@ -24,6 +24,10 @@
     "axios": "^0.21.1",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-graphql": "^0.12.0",
+    "graphql": "^15.5.0",
+    "graphql-import": "^1.0.2",
+    "graphql-tag": "^2.11.0",
     "redis": "^3.0.2",
     "socket.io": "3.0.4",
     "typescript": "^3.9.3"
