+++ api/src/gql/resolvers/Query/hello.ts
@@ -1,0 +1,3 @@
+export const hello = (): string => {
+  return 'Howdy'
+}
