+++ api/src/gql/resolvers/Query/index.ts
@@ -1,0 +1,1 @@
+export { hello } from './hello'
