+++ api/src/gql/resolvers/index.ts
@@ -1,0 +1,5 @@
+import * as Query from './Query'
+
+export const resolvers = {
+  hello: Query.hello,
+}
