+++ api/src/gql/schema.gql
@@ -1,0 +1,3 @@
+type Query {
+  hello: String
+}
